requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
python-dotenv==1.0.0
aiohttp==3.9.1
openai>=1.0.0
